+++ src/EMarker.cc
@@ -26,7 +26,7 @@
 
 using namespace degate;
 
-EMarker::EMarker(int _x, int _y, diameter_t _diameter) :
+EMarker::EMarker(float _x, float _y, diameter_t _diameter) :
 	Circle(_x, _y, _diameter)
 {
 }
@@ -81,25 +81,25 @@ void EMarker::print(std::ostream& os, int n_tabs) const
 		<< std::endl;;
 }
 
-void EMarker::shift_x(int delta_x)
+void EMarker::shift_x(float delta_x)
 {
 	Circle::shift_x(delta_x);
 	notify_shape_change();
 }
 
-void EMarker::shift_y(int delta_y)
+void EMarker::shift_y(float delta_y)
 {
 	Circle::shift_y(delta_y);
 	notify_shape_change();
 }
 
-void EMarker::set_x(int x)
+void EMarker::set_x(float x)
 {
 	Circle::set_x(x);
 	notify_shape_change();
 }
 
-void EMarker::set_y(int y)
+void EMarker::set_y(float y)
 {
 	Circle::set_y(y);
 	notify_shape_change();
